setup mvc

+++ src/main/java/uk/movie/LikeService/services/LikeService.java
@@ -1,14 +1,20 @@
 package uk.movie.LikeService.services;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import uk.movie.LikeService.entities.Like;
 import uk.movie.LikeService.repositories.LikeRepository;
 
+import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class LikeService {
     private final LikeRepository likeRepository;
 
@@ -20,19 +26,24 @@ public class LikeService {
         return likeRepository.findLikeByUserId(userId);
     }
 
+    @Transactional
     public void addOrRemoveLike(int userId, int movieId){
-        likeRepository.findLikeByMovieIdAndUserId(
-                movieId,userId
-        ).ifPresentOrElse((like -> {
+        //Send like entity to kafka topic as a json object
+        Optional<Like> likeOptional = likeRepository
+                .findLikeByMovieIdAndUserId(movieId,userId);
+        if(likeOptional.isPresent()){
+            Like like = likeOptional.get();
+            log.info("Found like to remove {}", like);
             likeRepository.delete(like);
+        }else{
             //Send like entity to kafka topic as a json object
-        }), ()->{
-            Like like = Like.builder()
+            log.info("Adding new like");
+            Like newLike = Like.builder()
                     .movieId(movieId)
                     .userId(userId)
+                    .createdAt(Timestamp.from(Instant.now()))
                     .build();
-            likeRepository.save(like);
-            //Send like entity to kafka topic as a json object
-        });
+            likeRepository.save(newLike);
+        }
     }
 }
